arrange UI vertically on narrow screens
Our UI is two boxes: an editor and a repl. Together they take up the
whole screen.

We've been developing the UI on our laptops, which have wide screens,
and it looks good to arrange the boxes horizontally there:

    |-----------------------|
    | editor    | repl      |
    |           |           |
    |           |           |
    |           |           |
    |-----------------------|

On narrow screens, like my phone, it would look better to arrange them
vertically:

    |--------|
    | editor |
    |        |
    |        |
    |--------|
    | repl   |
    |        |
    |        |
    |--------|

The UI now automatically switches between horizontal and vertical
layouts: horizontal if the screen is wider than it is tall, vertical
otherwise.