fixed editor height bug in safari
The reason this was happening is that we were inconsistent in how we
were setting the background color between left/right sides. On the left,
we set it on the textarea. On the right, we set it on the body div. This
is fine (though I disliked the inconsistency so I changed it) as long as
the editor textarea and the body div are the same height. That was the
case in chrome/firefox but not safari.

This stackoverflow answer sort of explains:
https://stackoverflow.com/a/33644245

In summary, I'm not 100% why this is happening but here's my best guess:

When you have a child element with percent height within a parent that
has an indefinite height, the percent cannot be calculated so it uses
auto, which essentially uses the default element height. And it looks
like flex-basis: auto means an element has indefinite height, so the
textarea was trying to use height: 100% and finding its parent had
indefinite height, so it defaulted to height: auto. So the fix was to
just remove the offending `height: 100%` which made the textarea
calculate its height at "100%" (I believe) by virtue of being the only
item in a flexbox.

The confusing part is that chrome/firefox worked. Actually, desktop
chrome/firefox worked, iOS chrome/firefox both broke, which is sort of
interesting in its own right: I guess mobile chrome/firefox only serve
to sync your data/change your browser UI, they don't actually implement
a different browser. But anyways my best guess here is that
chrome/firefox have some fix to this issue and safari does not. That's
hinted at in the stackoverflow link above, but the answer is a few years
old and from googling it seems like issues like this are actively
getting fixed.

Also safari had a default textarea margin of 1px, which was showing up
in a couple places, so in this commit I explicitly set it to 0.