reject multi-expression input in repl
But continue allowing it in the editor.

This is also how Chime's CLI behaves: files are expected to contain many
expressions, but REPL input is expected to only contain one.

I could be persuaded otherwise, but I submit that these examples behave
in surprising ways if we accept multi-expression input in the REPL:

   > set x 1
   undefined variable: x

   > set 'x 1
   1
   > x
   undefined variable: x

   > + 1 2
   2

^ All of these are instead parse errors if we are expecting a single
expression, which seems better.