print multi-line outputs on multiple lines
Straightforward enough.

Printing it all on one line made parse failures look particularly bad,
because the failures are pretty-printed and designed to have their
columns line up exactly. Like this:

    1:5:
      |
    1 | set x 1
      |     ^
    unexpected 'x'
    expecting end of input

I used the `pre` element to accomplish this. By default this element has
margin on top and bottom (at least in my browser). I removed the margin
on the top, but left it on the bottom, so we now have space between our
input/output pairs. This matches Chime's CLI behavior.