editor: "disable" example buttons that are no-ops
Now that the hello world example is also the first example in the
examples dropdown, the following somewhat-surprising experience is
possible:

  - load bel-repl.com for the first time
  - click on the examples drop down
  - click on the first example
  - ...nothing happens???

Since hello world is already loaded in the editor, nothing changes when
you click the hello world example.

To make this a bit more obvious, the examples dropdown now grays out the
currently-loaded example, and changes the pointer style to no longer
suggest that something will happen when you click it.

I also considered omitting the current example from the list, but didn't
like that as much.