quick hack to make examples prettier
When the examples load, the cursor position is at the end of the text.
Since the examples all end in parens, this triggers bracket matching.

eg:

    [(]cons 'hello 'world[)]

whereas this used to be cleaner:

    (cons 'hello 'world)

I'm not sure what a nice way would be to fix this. Maybe disabling
bracket matching when the editor does not have keyboard focus?

There is a really easy hack to fix it, though: add a space to the end of
all of the example code. This causes the cursor to be after a space when
the example loads, rather than after a paren, so no brackets are
highlighted. Much prettier.